requirements.txt
Pillow==9.5.0
streamlit==1.23.1
#tensorflow==2.12.0
tensorflow
tensorflow_intel
